requirements-eb.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
bcrypt==4.1.2
PyJWT==2.8.0
structlog==23.2.0